fix most compiler warnings

+++ ArchImpl/RISCV/RISCVArchSpecificImp.h
@@ -201,7 +201,7 @@ etiss::int32 RISCVArch::handleException(etiss::int32 cause, ETISS_CPU *cpu)
                 break;
             }
             etiss_uint32 irqLine = 0;
-            for (int i = 0; i < sizeof(mip_tmp) * 8; ++i)
+            for (size_t i = 0; i < sizeof(mip_tmp) * 8; ++i)
             {
                 // Highest interrupt line with highest priority
                 if (unlikely((mip_tmp >> i) & 0x1))
